requirements.txt
# Core testing framework
pytest>=7.0.0
pytest-html>=3.0.0
pytest-xdist>=3.0.0
allure-pytest>=2.10.0

# HTTP client and security testing
requests>=2.28.0
urllib3>=1.26.0
httpx>=0.24.0

# Security testing tools
sqlparse>=0.4.0
paramiko>=3.0.0

# JWT and crypto
PyJWT>=2.6.0
cryptography>=3.4.0

# Data handling and validation
pydantic>=1.10.0
jsonschema>=4.0.0
faker>=18.0.0

# Configuration and environment
pyyaml>=6.0.0
python-dotenv>=0.19.0

# Reporting and logging
colorlog>=6.0.0
tabulate>=0.9.0

# AI/ML for intelligent testing (optional)
openai>=1.0.0
anthropic>=0.7.0

# Performance testing (optional)
locust>=2.0.0

# Docker support (optional)
docker>=6.0.0
